+++ test/container.test.js
@@ -105,15 +105,31 @@ QUnit.test('should identify known types', function(assert) {
   ['mp3', 'aac', 'flac', 'ac3'].forEach(function(type) {
     const dataWithId3 = concatTypedArrays(id3Data, testData[type]);
     const dataWithId3Footer = concatTypedArrays(id3DataWithFooter, testData[type]);
-    const multipleId3 = concatTypedArrays(id3Data, id3Data, id3Data, testData[type]);
-    const multipleId3Footer = concatTypedArrays(id3DataWithFooter, id3DataWithFooter, id3DataWithFooter, testData[type]);
-    const multipleId3Mixed = concatTypedArrays(id3DataWithFooter, id3Data, id3DataWithFooter, id3Data, testData[type]);
+    const recursiveDataWithId3 = new Uint8Array([]
+      .concat(id3Data)
+      .concat(id3Data)
+      .concat(id3Data)
+      .concat(testData[type]));
+    const recursiveDataWithId3Footer = new Uint8Array([]
+      .concat(id3DataWithFooter)
+      .concat(id3DataWithFooter)
+      .concat(id3DataWithFooter)
+      .concat(testData[type]));
+
+    const differentId3Sections = new Uint8Array([]
+      .concat(id3DataWithFooter)
+      .concat(id3Data)
+      .concat(id3DataWithFooter)
+      .concat(id3Data)
+      .concat(testData[type]));
 
     assert.equal(detectContainerForBytes(dataWithId3), type, `id3 skipped and ${type} detected`);
     assert.equal(detectContainerForBytes(dataWithId3Footer), type, `id3 + footer skipped and ${type} detected`);
-    assert.equal(detectContainerForBytes(multipleId3), type, `multiple id3 skipped and ${type} detected`);
-    assert.equal(detectContainerForBytes(multipleId3Footer), type, `multiple id3 + footer skipped and ${type} detected`);
-    assert.equal(detectContainerForBytes(multipleId3Mixed), type, `multiple id3 with w/o footer skipped and ${type} detected`);
+
+    assert.equal(detectContainerForBytes(recursiveDataWithId3), type, `id3 x3 skipped and ${type} detected`);
+    assert.equal(detectContainerForBytes(recursiveDataWithId3Footer), type, `id3 + footer x3 skipped and ${type} detected`);
+    assert.equal(detectContainerForBytes(differentId3Sections), type, `id3 with/without footer skipped and ${type} detected`);
+
   });
 
   const notTs = concatTypedArrays(testData.ts, filler(188));